move the unaligned vector test to their own module
